+++ site/src/components/HelpTooltip/HelpTooltip.tsx
@@ -8,6 +8,7 @@ import {
 import HelpIcon from "@mui/icons-material/HelpOutline";
 import OpenInNewIcon from "@mui/icons-material/OpenInNew";
 import Link from "@mui/material/Link";
+import { Stack } from "components/Stack/Stack";
 import {
 	Popover,
 	PopoverContent,
@@ -15,8 +16,7 @@ import {
 	type PopoverProps,
 	PopoverTrigger,
 	usePopover,
-} from "components/Popover/Popover";
-import { Stack } from "components/Stack/Stack";
+} from "components/deprecated/Popover/Popover";
 import {
 	type FC,
 	type HTMLAttributes,
