+++ site/src/modules/dashboard/Navbar/UserDropdown/UserDropdownContent.tsx
@@ -17,8 +17,8 @@ import Tooltip from "@mui/material/Tooltip";
 import type * as TypesGen from "api/typesGenerated";
 import { CopyButton } from "components/CopyButton/CopyButton";
 import { ExternalImage } from "components/ExternalImage/ExternalImage";
-import { usePopover } from "components/Popover/Popover";
 import { Stack } from "components/Stack/Stack";
+import { usePopover } from "components/deprecated/Popover/Popover";
 import type { FC } from "react";
 import { Link } from "react-router-dom";
