+++ site/src/pages/ManagementSettingsPage/IdpSyncPage/IdpPillList.tsx
@@ -5,7 +5,7 @@ import {
 	Popover,
 	PopoverContent,
 	PopoverTrigger,
-} from "components/Popover/Popover";
+} from "components/deprecated/Popover/Popover";
 import type { FC } from "react";
 import { isUUID } from "utils/uuid";
