+++ site/src/pages/ManagementSettingsPage/UserTable/EditRolesButton.tsx
@@ -12,12 +12,12 @@ import {
 	HelpTooltipTrigger,
 } from "components/HelpTooltip/HelpTooltip";
 import { EditSquare } from "components/Icons/EditSquare";
+import { Stack } from "components/Stack/Stack";
 import {
 	Popover,
 	PopoverContent,
 	PopoverTrigger,
-} from "components/Popover/Popover";
-import { Stack } from "components/Stack/Stack";
+} from "components/deprecated/Popover/Popover";
 import { type ClassName, useClassName } from "hooks/useClassName";
 import type { FC } from "react";
