+++ site/src/pages/ManagementSettingsPage/UserTable/UserRoleCell.tsx
@@ -23,7 +23,7 @@ import {
 	Popover,
 	PopoverContent,
 	PopoverTrigger,
-} from "components/Popover/Popover";
+} from "components/deprecated/Popover/Popover";
 import type { FC } from "react";
 import { EditRolesButton } from "./EditRolesButton";
