+++ site/src/pages/TemplateVersionEditorPage/ProvisionerTagsPopover.tsx
@@ -6,12 +6,12 @@ import TextField from "@mui/material/TextField";
 import useTheme from "@mui/system/useTheme";
 import { FormFields, FormSection, VerticalForm } from "components/Form/Form";
 import { TopbarButton } from "components/FullPageLayout/Topbar";
+import { Stack } from "components/Stack/Stack";
 import {
 	Popover,
 	PopoverContent,
 	PopoverTrigger,
-} from "components/Popover/Popover";
-import { Stack } from "components/Stack/Stack";
+} from "components/deprecated/Popover/Popover";
 import { useFormik } from "formik";
 import { ProvisionerTag } from "modules/provisioners/ProvisionerTag";
 import { type FC, Fragment } from "react";
